Revert "Refactor: Complete application rewrite with GSAP and UI overhaul"
