--- jules-scratch/verification/error_page.html
@@ -1,1 +1,0 @@
-<html><head><meta name="color-scheme" content="light dark"></head><body><pre style="word-wrap: break-word; white-space: pre-wrap;">Internal Server Error</pre></body></html>
