+++ package.json
@@ -3,28 +3,26 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev --turbopack",
     "build": "next build",
     "start": "next start",
     "lint": "next lint"
   },
   "dependencies": {
     "@types/file-saver": "^2.0.7",
-    "@types/gsap": "^1.20.2",
     "file-saver": "^2.0.5",
-    "gsap": "^3.13.0",
     "html-to-image": "^1.11.13",
-    "lucide-react": "^0.539.0",
-    "next": "^15.4.6",
+    "lucide-react": "^0.509.0",
+    "next": "15.3.2",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
   },
   "devDependencies": {
-    "@tailwindcss/postcss": "^4.1.11",
-    "@types/node": "^24.2.1",
+    "@tailwindcss/postcss": "^4",
+    "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
-    "tailwindcss": "^4.1.11",
-    "typescript": "^5.9.2"
+    "tailwindcss": "^4",
+    "typescript": "^5"
   }
 }
