--- server.log
@@ -1,3 +1,0 @@
-
-> tierzen@0.1.0 dev
-> next dev
